Add new ZD strategies from literature, fix a bug, and various typos / reorg
