+++ axelrod/tests/strategies/test_meta.py
@@ -229,8 +229,8 @@ class TestMetaHunter(TestMetaPlayer):
         self.versus_test(opponent=axelrod.CyclerCCCD(),
                          expected_actions=actions)
 
-class TestMetaHunterAggressive(TestMetaPlayer):
 
+class TestMetaHunterAggressive(TestMetaPlayer):
     name = "Meta Hunter Aggressive"
     player = axelrod.MetaHunterAggressive
     expected_classifier = {
@@ -294,7 +294,7 @@ class TestMetaMajorityMemoryOne(TestMetaPlayer):
     }
 
     def test_strategy(self):
-        actions = [(C, C), (C, D), (D, C), (C, D), (D, C)]
+        actions = [(C, C), (C, D), (C, C), (C, D), (D, C)]
         self.versus_test(opponent=axelrod.Alternator(),
                          expected_actions=actions)
